Improve error propogation in filterImageFromURL
Should fix an issue where filterImageFromURL would silently fail when
the provided URL was pointing to something other than image. The Jimp
library would throw an error in this case, but the error was not
propogated to the original callers of the filterImageFromURL function
and the returned promise would never fulfill.